+++ content/src/controller/utils.ts
@@ -1,4 +1,7 @@
 import { InvalidRequestError, Pagination } from '../types'
+import { EntityContentItemReference } from '@dcl/hashing'
+import { Entity } from '@dcl/schemas'
+import { EntityField } from './Controller'
 
 export function paginationObject(url: URL, maxPageSize: number = 1000): Pagination {
   const pageSize = url.searchParams.has('pageSize') ? parseInt(url.searchParams.get('pageSize')!, 10) : 100
@@ -34,4 +37,21 @@ export function asEnumValue<T extends { [key: number]: string }>(
     const match = validEnumValues.has(stringToMap)
     return match ? (stringToMap as T[keyof T]) : 'unknown'
   }
+}
+
+export function maskEntity(fullEntity: Entity, fields?: EntityField[]): Entity {
+  const { id, type, timestamp, version } = fullEntity
+  let content: EntityContentItemReference[] = []
+  let metadata: any
+  let pointers: string[] = []
+  if ((!fields || fields.includes(EntityField.CONTENT)) && fullEntity.content) {
+    content = fullEntity.content
+  }
+  if (!fields || fields.includes(EntityField.METADATA)) {
+    metadata = fullEntity.metadata
+  }
+  if ((!fields || fields.includes(EntityField.POINTERS)) && fullEntity.pointers) {
+    pointers = fullEntity.pointers
+  }
+  return { version, id, type, timestamp, pointers, content, metadata }
 }
